Quantity is now shared
