--- src/Developurr.Orderly.Domain/Order/Validators/QuantityValidatorConfig.cs
@@ -1,7 +1,0 @@
-namespace Developurr.Orderly.Domain.Order.Validators;
-
-public static class QuantityValidatorConfig
-{
-    public const int QuantityMin = 0;
-    public const int QuantityMax = int.MaxValue-1;
-}
