+++ tests/UnitTests/Developurr.Orderly.Domain.UnitTests/Order/ValueObjects/QuantityTest.cs
@@ -1,5 +1,6 @@
 using Developurr.Orderly.Domain.Exceptions;
 using Developurr.Orderly.Domain.Order.ValueObjects;
+using Developurr.Orderly.Domain.Shared.ValueObjects;
 using Developurr.Orderly.Domain.UnitTests.TestUtils.Constants;
 
 namespace Developurr.Orderly.Domain.UnitTests.Order.ValueObjects;
