+++ tests/UnitTests/Developurr.Orderly.Domain.UnitTests/TestUtils/Quantity/QuantityFixture.cs
@@ -2,9 +2,9 @@ namespace Developurr.Orderly.Domain.UnitTests.TestUtils.Quantity;
 
 public sealed class QuantityFixture
 {
-    public static Developurr.Orderly.Domain.Order.ValueObjects.Quantity CreateQuantity()
+    public static Shared.ValueObjects.Quantity CreateQuantity()
     {
-        return Developurr.Orderly.Domain.Order.ValueObjects.Quantity.Create(
+        return Shared.ValueObjects.Quantity.Create(
             Constants.Constants.Quantity.Value    
         );
     }
